% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{args}
\alias{args}
\title{Construct query arguments}
\usage{
args(...)
}
\arguments{
\item{...}{The R epxressions that represents the arguments to the query.
Each expression must be of the form \code{variable <- value},
which will cause \code{value} to be bound to \code{variable} in the query}
}
\value{
A list of the epxressions parsed in a way that is suitable
  for inclusion in a query.
}
\description{
This function constructs arguments for a query starting
from a series of R expressions
}
