% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{c_query}
\alias{c_query}
\title{Combine queries}
\usage{
c_query(a, b)
}
\arguments{
\item{a}{A query as returned by \code{\link{query}}}

\item{b}{A query as returned by \code{\link{query}}}
}
\value{
Returns the fields of the \code{b} query appended to the corresponding fields
  of the \code{a} query
}
\description{
This function combines queries by concatenating their \code{find}, \code{where}
and \code{args} fields
}
