% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{convert_pull_query_results}
\alias{convert_pull_query_results}
\title{Converts the results of a pull expression query}
\usage{
convert_pull_query_results(query.res, flatten = TRUE,
  exclude.db.ids = TRUE)
}
\arguments{
\item{query.res}{A \code{list} of \code{data.frames}. The results of the query}

\item{flatten}{If \code{TRUE} nested \code{data.frames} are recursively flattened
using \code{wick::flatten_data_frame}}

\item{exclude.db.ids}{If \code{TRUE}, columns representing db identifiers are removed
from the result}
}
\value{
Returns a \code{data.frame} of converted query results
}
\description{
Converts the results of a pull expression query
}
