% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{convert_query_results}
\alias{convert_query_results}
\title{Converts query results}
\usage{
convert_query_results(query, x, simplify = TRUE)
}
\arguments{
\item{query}{The query that generated this result}

\item{x}{The query result}

\item{simplify}{If \code{TRUE} a \code{data.frame} with a single column
is simplified to a vector}
}
\value{
Returns the query results in \code{data.frame} format. Column names are derived
  from the \code{find} clause of the query
}
\description{
This function converts query results so that they can be more easily consumed by R.
The data is transformed into a \code{data.frame}, and variables are converted
to \code{numeric} when possible.
}
