% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{d}
\alias{d}
\title{Convert expressions to Datalog}
\usage{
d(...)
}
\arguments{
\item{...}{The epxressions to convert to Datalog}
}
\value{
Returns a list of strings where each element
  corresponds to the string representation of the corresponding expression
  in \code{...}
}
\description{
This function converts R expressions to Datalog. The following rules apply
when converting expressions to Datalog:
\itemize{
  \item Infix operators are converted to prefix
  \item Function calls are converted to S-expressions (e.g. max(v) is converted to (max v))
  \item The \code{.} symbol is converted to \code{_}
  \item String containing the \code{/} character are assumed to represent
    qualified symbol names and are converted by pre-prending the \code{:} character
  \item Names prefixed with \code{!!} are assumed to represent R symbols and are substituted
    with their bound value
}
Note the following:
\itemize{
  \item This function only performs syntax conversion and does not check the
    correctness of the resulting expressions
  \item For the same reason, even though you may use function names that exist in R
    (e.g. \code{max}), this function is not actually invoking the corresponding R
    function at all, therefore you should not assume that you can pass it the same arguments
    that you would in R. For instance the expression \code{max(v, is.na = TRUE)} would make no sense
    since the R \code{max} function is not invoked at any point during the conversion process
}
}
