% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ensure_server_url}
\alias{ensure_server_url}
\title{Ensure server.url has been set correctly}
\usage{
ensure_server_url(server.url)
}
\arguments{
\item{server.url}{The server URL, possibly \code{NULL}}
}
\value{
Returns \code{server.url} if not \code{NULL}. Otherwise returns the global
  \code{pkg.env$server.url} value
}
\description{
Ensure server.url has been set correctly
}
