% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{flatten_data_frame}
\alias{flatten_data_frame}
\title{Flattens a list of data frames}
\usage{
flatten_data_frame(x)
}
\arguments{
\item{x}{A list of data frames}
}
\value{
Returns a \code{data.fame}
}
\description{
This function flattens a list of data frames
using \code{jsonlite::flatten}
}
