% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{generate_or}
\alias{generate_or}
\title{Generate multiple or clauses for several alternatives}
\usage{
generate_or(alternatives, ...)
}
\arguments{
\item{alternatives}{A character vector of alternatives}

\item{...}{The partial clause}
}
\value{
Returns an \code{or} expression suitable for inclusion in a query
}
\description{
This function generates multiple \code{or} clauses by combining a partial clause with multiple
alternatives. For instance given the alternatives \code{c("A", "B")} and the partial clause
\code{?p, epitope/name} this function generates the or clause
\code{
  or(
    d(?p, epitope/name, "A"),
    d(?p, epitope/name, "B")
  )
}
}
\examples{

generate_or(
  alternatives = c("A", "B"), ?p, epitope/name
)

}
