% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{is_pull_query}
\alias{is_pull_query}
\title{Checks whether a query contains a pull expression}
\usage{
is_pull_query(query)
}
\value{
Returns a boolean indicating whether this query contains a pull expression
  or not
}
\description{
Checks whether a query contains a pull expression
}
