% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{or}
\alias{or}
\title{Construct or clauses}
\usage{
or(...)
}
\arguments{
\item{...}{The individual clauses, constructed with \code{\link{d}}}
}
\value{
The expressions parsed in a way that is suitable
  for inclusion in a query
}
\description{
Construct or clauses
}
\examples{

qq <- query(find(?e),
    where(
        or(
           d(?e, gene/hugo),
           d(?e, variant/coordinates)
        )
    )
)


}
