% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse an individual query argument}
\usage{
parse_arg(s)
}
\arguments{
\item{s}{The expression to parse}
}
\value{
Returns the expression parsed into a named list with two elements
  named \code{in} and \code{args}
}
\description{
Parse an individual query argument
}
