% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{parse_pull_data_literals}
\alias{parse_pull_data_literals}
\title{Parses data literals in a pull expression}
\usage{
parse_pull_data_literals(s)
}
\arguments{
\item{s}{An unevaluated R expression to be deparsed and processed}
}
\value{
A string representation of the expression, with the appropriate
  substitutions performed
}
\description{
This function parses the data literals in a pull expression
(\code{c()} and \code{{key = val}})
and converts them to the appropriate R function calls
}
