% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{query}
\alias{query}
\title{Construct a Datalog query}
\usage{
query(...)
}
\arguments{
\item{...}{The components of the query, these should be invocations
of the \code{\link{find}}, \code{\link{where}} and optionally \code{\link{args}}
functions}
}
\value{
Returns the Datalog query in list representation
}
\description{
Constructus a Datalog query
}
