% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{set_server_url}
\alias{set_server_url}
\title{Globally set the server URL}
\usage{
set_server_url(server.url)
}
\arguments{
\item{server.url}{The server URL}
}
\value{
Returns the previous \code{server.url} value, invisibly
}
\description{
This function sets the server URL globally for the entire package (so that you
  don't have to pass around a \code{server.url} argument with every query)
}
