% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{try_numeric}
\alias{try_numeric}
\title{Attempt conversion to numeric}
\usage{
try_numeric(s)
}
\arguments{
\item{s}{The string to convert}
}
\value{
Returns \code{as.numeric(s)}, os \code{s} if numeric conversion fails
}
\description{
Attempt conversion to numeric
}
