% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{where}
\alias{where}
\title{Construct the where portion of a Datalog query}
\usage{
where(...)
}
\arguments{
\item{...}{R expressions to be converted using \code{\link{d}}}
}
\value{
A list of parsed expressions, suitable for inlcusion in a
  \code{\link{query}}
}
\description{
Construct the where portion of a Datalog query
}
