% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalog.R
\name{with}
\alias{with}
\title{Construct the with portion of a Datalog query}
\usage{
with(...)
}
\arguments{
\item{...}{R expressions to be converted using \code{\link{d}}}
}
\value{
A list of parsed expressions, suitable for inlcusion in a
  \code{\link{query}}
}
\description{
Construct the with portion of a Datalog query
}
